Add files via upload
In this project, we navigated the process of chip design by utilizing a range of hardware design principles, tools, and methodologies. The focus was on transforming high-level logic descriptions into a physical chip design prepared for manufacturing and commercial release.

The project effectively showcased the complete development of a chip, starting from high-level language descriptions and culminating in physical hardware. By leveraging HDL tools, the RISC-V architecture, and cutting-edge manufacturing techniques, we achieved a reliable and efficient design ready for deployment